Complete 1/3 IIFE wrapped factory functions; for gameboard object array
Initial approach was unsophisticated and WET, meaning lots of repitition
to create the 9 item object array which would form the basis for the
displayed 3x3 game grid. Ended figuring out how to automate using a
'while' loop to create array for the object 'keys', nested inside of
a 'for' loop which created the values (gameboard[index]) of the key:value
pairs. The resulting pairs were then pushed to the original gameboard
array created outside both loops but inside the IIFE function. The IIFE's
inner function is then used to return the mutated 'gameBoard' array.
Calling the ' const = function ' outside of the (IIFE) would return the
innermost function, and give you the gameboard array.

I know. I ton of work for this tiny return. But this was good practice
for this kind of thinking. Although I need to get a lot better at
explaining this type of stuff in writing.