Logic to create and add objects to array; refactor event handling
Previously, only display/DOM was being changed to reflect the player
moves. Added logic to create new Objects from the PlayMove class,
and then push those objects into an open-ended array, gameBoardArray.

Refactored event handling so that .addEventListener wasn't in the
constructor function and being called immediately each time, b/c
there was an issue with the alert messages from playerOneChoice and
playerTwoChoice being called as soon as page loaded, without waiting
for the 'click' event to happen.

At This Point: Clicking the Player 1/Player 2 button triggers alert
message showing current player and marker selection. Clicking on any
of the squares on the gameboard results in the selected marker being
displayed on the UI, the position incremented to reflect the board
tile that was clicked on, and the the resulting marker/position combo
being stored as object and pushed into the gameBoardArray.