Add new methods: turn and final check win; create empty game board
Fix error in previous program logic where new marker objects were
simply being appendeded to the gameBoardArray. This effectively
created an unordered array, when we want a ordered array so as to
have index references to compare for win condition. This is why the
"createBoard" method was necessary to add.

Created separate methods for checking the win condition after each
turn, and also at the end of the round (when length of the gameboard
array was > 9). Both win condition methods work as intended.

Still NTD:
1. Add DOM element to display current player/marker selection
2. Add DOM element to display Win Message, i/o of current alert
3. Format Grid Tiles and Markers so they look nice and pretty